requirements.txt
streamlit>=1.28.0
anthropic>=0.25.0
pypdf>=3.17.0
python-docx>=1.1.0
reportlab>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system == "Windows"
